syn: an attempt at fixing CHECK
This isn't quite right but I can't find a way to recurse in sublime-syntax ($self is accepted but doesn't work).

Because I can't _refine_ something that was already captured, I can't do a straight-forward translation of the textmate grammar here so I'm sure this is, disregarding the $self thing, still different in noticeable ways.
We'll call it good enough for now though.